Move Lambda unit tests into subdirectory
