+++ tests/unit/services/lambda_/test_lambda_legacy.py
